+++ packages/spacecat-shared-data-access/test/unit/service/index.test.js
@@ -93,6 +93,12 @@ describe('Data Access Object Tests', () => {
     'getImportUrlsByJobIdAndStatus',
   ];
 
+  const experimentFunctions = [
+    'getExperiments',
+    'upsertExperiment',
+    'getExperiment',
+  ];
+
   let dao;
 
   before(() => {
@@ -151,7 +157,8 @@ describe('Data Access Object Tests', () => {
       ...configurationFunctions,
       ...siteTopPagesFunctions,
       ...importJobFunctions,
-      ...importUrlFunctions]);
+      ...importUrlFunctions,
+      ...experimentFunctions]);
     Object.keys(dao).forEach((funcName) => {
       expect(expectedFunctions).to.include(funcName);
     });